+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: abackman <abackman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/05 12:42:34 by abackman          #+#    #+#             */
-/*   Updated: 2021/11/05 13:15:48 by abackman         ###   ########.fr       */
+/*   Updated: 2021/11/05 13:35:26 by abackman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 char	*ft_strchr(const char *s, int c)
 {
 	char	*p;
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = ft_strlen(s);
